+++ temba/archives/views.py
@@ -4,7 +4,7 @@ from smartmin.views import SmartCRUDL, SmartListView, SmartReadView
 
 from django.http import HttpResponseRedirect
 
-from temba.orgs.views import OrgObjPermsMixin, OrgPermsMixin
+from temba.orgs.mixins import OrgObjPermsMixin, OrgPermsMixin
 from temba.utils.views import SpaMixin
 
 from .models import Archive
