+++ temba/contacts/views.py
@@ -33,16 +33,9 @@ from temba.archives.models import Archive
 from temba.channels.models import Channel
 from temba.mailroom.events import Event
 from temba.notifications.views import NotificationTargetMixin
+from temba.orgs.mixins import OrgObjPermsMixin, OrgPermsMixin
 from temba.orgs.models import User
-from temba.orgs.views import (
-    BaseExportView,
-    DependencyDeleteModal,
-    DependencyUsagesModal,
-    MenuMixin,
-    ModalMixin,
-    OrgObjPermsMixin,
-    OrgPermsMixin,
-)
+from temba.orgs.views import BaseExportView, DependencyDeleteModal, DependencyUsagesModal, MenuMixin, ModalMixin
 from temba.tickets.models import Ticket, Topic
 from temba.utils import json, on_transaction_commit
 from temba.utils.dates import datetime_to_timestamp, timestamp_to_datetime
