+++ temba/msgs/views.py
@@ -28,17 +28,9 @@ from django.views.generic import RedirectView
 from temba import mailroom
 from temba.archives.models import Archive
 from temba.mailroom.client.types import Exclusions
+from temba.orgs.mixins import OrgFilterMixin, OrgObjPermsMixin, OrgPermsMixin
 from temba.orgs.models import Org
-from temba.orgs.views import (
-    BaseExportView,
-    DependencyDeleteModal,
-    DependencyUsagesModal,
-    MenuMixin,
-    ModalMixin,
-    OrgFilterMixin,
-    OrgObjPermsMixin,
-    OrgPermsMixin,
-)
+from temba.orgs.views import BaseExportView, DependencyDeleteModal, DependencyUsagesModal, MenuMixin, ModalMixin
 from temba.schedules.views import ScheduleFormMixin
 from temba.templates.models import Template, TemplateTranslation
 from temba.utils import json, languages
